Fix/event store (#708)
* Fix take.

* Better tests.

* Fix build.

* Another fix.

* Reverted fix.
# Conflicts:
#	backend/tests/Squidex.Infrastructure.Tests/Collections/ImmutableListTests.cs

+++ backend/src/Squidex.Infrastructure/EventSourcing/IEventStore.cs
@@ -19,9 +19,9 @@ namespace Squidex.Infrastructure.EventSourcing
 
         Task<IReadOnlyList<StoredEvent>> QueryAsync(string streamName, long streamPosition = 0);
 
-        IAsyncEnumerable<StoredEvent> QueryAllReverseAsync(string? streamFilter = null, Instant timestamp = default, long take = long.MaxValue, CancellationToken ct = default);
+        IAsyncEnumerable<StoredEvent> QueryAllReverseAsync(string? streamFilter = null, Instant timestamp = default, int take = int.MaxValue, CancellationToken ct = default);
 
-        IAsyncEnumerable<StoredEvent> QueryAllAsync(string? streamFilter = null, string? position = null, long take = long.MaxValue, CancellationToken ct = default);
+        IAsyncEnumerable<StoredEvent> QueryAllAsync(string? streamFilter = null, string? position = null, int take = int.MaxValue, CancellationToken ct = default);
 
         Task AppendAsync(Guid commitId, string streamName, ICollection<EventData> events);
